# Conflicts:
#	api-demo/src/main/java/com/jk/model/Tree.java

+++ api-demo/src/main/java/com/jk/model/Book.java
@@ -46,7 +46,7 @@ public class Book implements Serializable {
 
     private Integer bookCount;
 
-    private Double hits;
+    private Integer hits;
 
     private Integer bookStar;
 
@@ -63,4 +63,6 @@ public class Book implements Serializable {
     private Integer buyCount;
 
     private Integer bookTypePid;
+
+
 }
